extract lifespan and add functionality to write swagger file
