+++ apps/ls-api/services/environment.py
@@ -1,0 +1,5 @@
+import os
+
+DEVELOPMENT_MODE = "development"
+
+mode = os.getenv("ENVIRONMENT", DEVELOPMENT_MODE)
