+++ packages/api-client/package.json
@@ -9,7 +9,7 @@
     "lint": "eslint .",
     "format": "eslint . --fix",
     "check-types": "tsc --noEmit",
-    "codegen:lgg": "rm -rf src/lgg/generated && openapi-generator-cli generate -i http://localhost:8000/openapi.json -g typescript-fetch -o src/lgg/generated --additional-properties=disallowAdditionalPropertiesIfNotPresent=false,supportsES6=true,useSingleRequestParameter=false --enable-post-process-file --remove-operation-id-prefix",
+    "codegen:lgg": "rm -rf src/lgg/generated && openapi-generator-cli generate -i ../../apps/ls-api/docs/openapi.json -g typescript-fetch -o src/lgg/generated --additional-properties=disallowAdditionalPropertiesIfNotPresent=false,supportsES6=true,useSingleRequestParameter=false --enable-post-process-file --remove-operation-id-prefix",
     "codegen": "pnpm codegen:lgg"
   },
   "devDependencies": {
